requirements.txt
beautifulsoup4==4.13.4
chromadb==1.3.0
google-generativeai==0.8.5
langchain==1.0.3
langchain-chroma==1.0.0
langchain-community==0.4.1
langchain-core==1.0.2
langchain-google-genai==3.0.0
langchain-huggingface==1.0.0
langgraph==1.0.2
langgraph-prebuilt==1.0.2
pydantic==2.10.6
pypdf==4.3.1
python-docx==1.2.0
requests==2.32.5
sentence-transformers==4.1.0
transformers==4.53.0
tqdm==4.67.1
